Moved datasets.mock to io.mock.artificial. Closes #93.
