+++ phy/cluster/manual/tests/test_utils.py
@@ -13,7 +13,7 @@ from numpy.testing import assert_array_equal as ae
 from pytest import raises
 
 from .._utils import _unique, _spikes_in_clusters
-from ....datasets.mock import artificial_spike_clusters
+from ....io.mock.artificial import artificial_spike_clusters
 
 
 #------------------------------------------------------------------------------
