+++ phy/plot/tests/test_waveforms.py
@@ -13,8 +13,8 @@ from ...cluster.manual.cluster_metadata import ClusterMetadata
 from ...utils.logging import set_level
 from ..waveforms import Waveforms, WaveformView
 from ...utils._color import _random_color
-from ...datasets.mock import (artificial_waveforms, artificial_masks,
-                              artificial_spike_clusters)
+from ...io.mock.artificial import (artificial_waveforms, artificial_masks,
+                                   artificial_spike_clusters)
 from ...electrode.mea import staggered_positions
 from ...utils.testing import show_test
