+++ phy/waveform/tests/test_loader.py
@@ -14,7 +14,7 @@ from numpy.testing import assert_array_equal as ae
 import numpy.random as npr
 from pytest import raises
 
-from ...datasets.mock import artificial_traces
+from ...io.mock.artificial import artificial_traces
 from ..loader import _slice, WaveformLoader
